requirements.txt
asgiref==3.2.10
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.2
channels==2.4.0
click==7.1.2
colorama==0.4.3
constantly==15.1.0
cryptography==3.0
daphne==2.5.0
dj-database-url==0.5.0
Django==3.0.8
django-bmemcached==0.3.0
django-cors-headers==3.4.0
django-heroku-memcacheify==1.0.0
django-pylibmc==0.6.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
docopt==0.6.2
flake8==3.8.3
glob2==0.7
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
iniconfig==1.0.1
jedi==0.17.2
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
parse==1.16.0
parse-type==0.5.2
parso==0.7.1
pathtools==0.1.2
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyHamcrest==2.0.2
pylibmc==1.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==6.0.1
pytest-asyncio==0.14.0
pytest-django==3.9.0
pytest-watch==4.2.0
python-binary-memcached==0.29.0
python-decouple==3.3
pytz==2020.1
sentry-sdk==0.16.2
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
uhashring==1.2
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
watchdog==0.10.3
websockets==8.1
whitenoise==5.1.0
wincertstore==0.2
zope.interface==5.1.0
